requirements.txt
pytest
radon
flake8
bandit
pip-audit
PyYAML
pandas
matplotlib
# mutmut
